working on final render

+++ src/common.hpp
@@ -11,7 +11,8 @@
 const double INFIN = std::numeric_limits<double>::infinity();
 const double PI = 3.1415926535897932385;
 
-inline float degreesToRadians(float degrees) { return degrees * PI / 180.0f; }
+inline float degreesToRadians(float degrees) { return degrees * (float)PI / 180.0f; }
+
 inline float randomFloat() {
     static std::uniform_real_distribution<float> distrubtion(0.0f, 1.0f);
     static std::mt19937 gen;
